Removed all logic from controllers. They are responsible for status code 2XX

+++ EventsMicroservice/Application/MediatR/Handlers/EventCommentHandlers/EventCommentDeleteCommandHandler.cs
@@ -2,6 +2,7 @@
 using Application.UnitOfWork.Interfaces;
 using AutoMapper;
 using MediatR;
+using ArgumentException = Shared.Exceptions.ServerExceptions.ArgumentException;
 
 namespace Application.MediatR.Handlers.EventCommentHandlers
 {
@@ -14,9 +15,14 @@ namespace Application.MediatR.Handlers.EventCommentHandlers
 
         public async Task Handle(EventCommentDeleteCommand request, CancellationToken cancellationToken)
         {
+            // TODO: Добавить проверку, что пользователь является создателем отзыва.
+
             var comment = await _unitOfWork.EventCommentRepository.GetByIdAsync(request.Id, cancellationToken);
             if (comment == null)
                 return;
+
+            if (request.RouteEventId != comment.EventId)
+                throw new ArgumentException("You are not allowed to delete this comment for the event.");
 
             await _unitOfWork.EventCommentRepository.DeleteAsync(comment, cancellationToken).ConfigureAwait(false);
         }
